init SendEmailVerificationTaskMixin: refactor SendEmailVerificationTask
