+++ english_diary/users/mixins/__init__.py
@@ -1,0 +1,1 @@
+from .tasks import SendEmailVerificationTaskMixin
